Added support for specifiying texture options
Textures can now be given a color space, min and mag filter options for
mipmaps as well as texture wrapping objects for S T and R coordinates.

All paramaters are optional.

Setting textures to be in the Linear colorspace rather than Gamma makes
them be stored as SRGB or SRGB_ALPHA formatted textures (SRGB_ALPHA if
transparency is supported).