Supporting a 64 bit build on Windows now
Refactored a bit of the CMake project to support an explicit directory for
the Bin directory of GLEW and GLFW. This makes it a bit more clear what
directories need to be pointed to where and when.

Apparantly in x64 builds on Windows std::vector returns size() as size_t
rather than unsigned int. Hopefully this doesn't mess with other
platforms. Maybe g++ was just auto-casting while MSVC likes to throw
warnings. size_t is now used for a bunch of internal size-related values